requirements.txt
streamlit
streamlit-aggrid
pandas
matplotlib
mysql-connector-python
plotly.express